pyproject.toml
[tool.poetry]
name = "xif"
version = "0.1.0"
description = "CLI Utility for Adding and Removing Metadata for AI Art"
license = "MIT"
authors = ["Brian Cohan <briancohan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
rasterio = "^1.3.9"
pillow = "^10.1.0"
exif = "^1.6.0"
bs4 = "^0.0.1"
typer = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
xif = "xif.cli:app"
